requirements.txt
anyio==4.3.0
certifi==2024.2.2
click==8.1.7
exceptiongroup==1.2.0
Faker==24.3.0
granian==1.2.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.6
importlib_metadata==7.1.0
litestar==2.7.0
markdown-it-py==3.0.0
mdurl==0.1.2
msgspec==0.18.6
multidict==6.0.5
polyfactory==2.15.0
Pygments==2.17.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
rich==13.7.1
rich-click==1.7.4
six==1.16.0
sniffio==1.3.1
typer==0.9.0
typing_extensions==4.10.0
uvloop==0.18.0
zipp==3.18.1
